fix: dont write to /ips-hosting folder
Instead we use the /tmp folder to store the steamcmd.
With this change all images are compatible with ReadonlyRootFs as long as a TmpFs is mounted.
Its also more performant to write to a tmpfs instead of writing to the container layer.
gmod raw content has also been moved from the container layer to the volume.
We dont use the /tmp folder in that case, becase /tmp is often an in-memory-fs and the raw content is multiple gb in size.